+++ cfgov/v1/atomic_elements/organisms.py
@@ -3,23 +3,21 @@ import json
 import requests
 
 from django import forms
-from django.template.loader import render_to_string
-from django.utils.functional import cached_property
-
 from django.apps import apps
+from django.template.loader import render_to_string
 from django.utils.encoding import smart_text
+from django.utils.functional import cached_property
+from jinja2 import Markup
 from wagtail.contrib.table_block.blocks import TableBlock, TableInput
 from wagtail.wagtailcore import blocks
+from wagtail.wagtaildocs.blocks import DocumentChooserBlock
 from wagtail.wagtailimages import blocks as images_blocks
 from wagtail.wagtailsnippets.blocks import SnippetChooserBlock
-from wagtail.wagtaildocs.blocks import DocumentChooserBlock
 
 from . import atoms, molecules
 from .. import blocks as v1_blocks
-from ..util import ref
 from ..models.snippets import Contact as ContactSnippetClass
-
-from jinja2 import Markup
+from ..util import ref
 
 
 class Well(blocks.StructBlock):
@@ -55,9 +53,12 @@ class ImageText5050Group(blocks.StructBlock):
 
 class ImageText2575Group(blocks.StructBlock):
     heading = blocks.CharBlock(icon='title', required=False)
-    should_link_image = blocks.BooleanBlock(default=False,
-                                            required=False,
-                                            help_text='Check this to link all images to the URL of the first link in their unit\'s list, if there is a link.')
+    should_link_image = blocks.BooleanBlock(
+        default=False,
+        required=False,
+        help_text=('Check this to link all images to the URL of the first '
+                   'link in their unit\'s list, if there is a link.')
+    )
     image_texts = blocks.ListBlock(molecules.ImageText2575())
 
     class Meta:
@@ -99,7 +100,9 @@ class EmailSignUp(blocks.StructBlock):
     text = blocks.CharBlock(required=False)
     gd_code = blocks.CharBlock(required=False)
 
-    form_field = blocks.ListBlock(molecules.FormFieldWithButton(), icon='mail', required=False)
+    form_field = blocks.ListBlock(molecules.FormFieldWithButton(),
+                                  icon='mail',
+                                  required=False)
 
     class Meta:
         icon = 'mail'
@@ -110,13 +113,31 @@ class EmailSignUp(blocks.StructBlock):
 
 
 class RegComment(blocks.StructBlock):
-    document_id = blocks.CharBlock(required=True, label='Document ID',
-                                   help_text='Federal Register document ID number to which the comment should be submitted. Should follow this format: CFPB-YYYY-####-####')
-    generic_regs_link = blocks.BooleanBlock(required=False, default=True,
-                                            label='Use generic Regs.gov link?',
-                                            help_text='If unchecked, the link to comment at Regulations.gov if you want to add attachments will link directly to the document given above. Leave this checked if this comment form is being published before the full document is live at Regulations.gov, then uncheck it when the full document has been published.')
-    id = blocks.CharBlock(required=False, label='Form ID',
-                          help_text='Sets the `id` attribute in the form\'s markup. If not set, the form will be assigned a base id of `o-reg-comment_` with a random number appended.')
+    document_id = blocks.CharBlock(
+        required=True,
+        label='Document ID',
+        help_text=('Federal Register document ID number to which the comment '
+                   'should be submitted. Should follow this format: '
+                   'CFPB-YYYY-####-####')
+    )
+    generic_regs_link = blocks.BooleanBlock(
+        required=False,
+        default=True,
+        label='Use generic Regs.gov link?',
+        help_text=('If unchecked, the link to comment at Regulations.gov if '
+                   'you want to add attachments will link directly to the '
+                   'document given above. Leave this checked if this comment '
+                   'form is being published before the full document is live '
+                   'at Regulations.gov, then uncheck it when the full '
+                   'document has been published.')
+    )
+    id = blocks.CharBlock(
+        required=False,
+        label='Form ID',
+        help_text=('Sets the `id` attribute in the form\'s markup. If not '
+                   'set, the form will be assigned a base id of '
+                   '`o-reg-comment_` with a random number appended.')
+    )
 
     class Meta:
         icon = 'form'
@@ -125,20 +146,41 @@ class RegComment(blocks.StructBlock):
 
 class RelatedPosts(blocks.StructBlock):
     limit = blocks.CharBlock(default='3', label='Limit')
-    show_heading = blocks.BooleanBlock(required=False, default=True,
-                                       label='Show Heading and Icon?',
-                                       help_text='This toggles the heading and'
-                                                 + ' icon for the related types.')
-    header_title = blocks.CharBlock(default='Further reading', label='Slug Title')
+    show_heading = blocks.BooleanBlock(
+        required=False,
+        default=True,
+        label='Show Heading and Icon?',
+        help_text=('This toggles the heading and '
+                   'icon for the related types.')
+    )
+    header_title = blocks.CharBlock(
+        default='Further reading',
+        label='Slug Title'
+    )
 
-    relate_posts = blocks.BooleanBlock(required=False, default=True,
-                                       label='Blog Posts', editable=False)
-    relate_newsroom = blocks.BooleanBlock(required=False, default=True,
-                                          label='Newsroom', editable=False)
-    relate_events = blocks.BooleanBlock(required=False, default=True,
-                                        label='Events')
+    relate_posts = blocks.BooleanBlock(
+        required=False,
+        default=True,
+        label='Blog Posts',
+        editable=False
+    )
+    relate_newsroom = blocks.BooleanBlock(
+        required=False,
+        default=True,
+        label='Newsroom',
+        editable=False
+    )
+    relate_events = blocks.BooleanBlock(
+        required=False,
+        default=True,
+        label='Events'
+    )
 
-    specific_categories = blocks.ListBlock(blocks.ChoiceBlock(choices=ref.related_posts_categories, required=False), required=False)
+    specific_categories = blocks.ListBlock(
+        blocks.ChoiceBlock(choices=ref.related_posts_categories,
+                           required=False),
+        required=False
+    )
 
     class Meta:
         icon = 'link'
@@ -174,12 +216,15 @@ class Table(blocks.StructBlock):
         template = '_includes/organisms/table.html'
         label = ' '
 
+
 class BureauStructurePosition(blocks.StructBlock):
     office_name = blocks.CharBlock()
     lead = v1_blocks.PlaceholderCharBlock(placeholder="Name")
     title = blocks.StructBlock([
-        ('line_1', v1_blocks.PlaceholderCharBlock(required=False, placeholder="Title 1")),
-        ('line_2', v1_blocks.PlaceholderCharBlock(required=False, placeholder="Title 2"))
+        ('line_1', v1_blocks.PlaceholderCharBlock(required=False,
+                                                  placeholder="Title 1")),
+        ('line_2', v1_blocks.PlaceholderCharBlock(required=False,
+                                                  placeholder="Title 2"))
     ])
 
 
@@ -187,8 +232,10 @@ class BureauStructureDivision(blocks.StructBlock):
     division = v1_blocks.PlaceholderCharBlock(label='Division')
     division_lead = v1_blocks.PlaceholderCharBlock(placeholder="Name")
     title = blocks.StructBlock([
-        ('line_1', v1_blocks.PlaceholderCharBlock(required=False, placeholder="Title 1")),
-        ('line_2', v1_blocks.PlaceholderCharBlock(required=False, placeholder="Title 2"))
+        ('line_1', v1_blocks.PlaceholderCharBlock(required=False,
+                                                  placeholder="Title 1")),
+        ('line_2', v1_blocks.PlaceholderCharBlock(required=False,
+                                                  placeholder="Title 2"))
     ])
     link_to_division_page = atoms.Hyperlink(required=False)
     offices = blocks.ListBlock(BureauStructurePosition(required=False))
@@ -203,8 +250,10 @@ class BureauStructure(blocks.StructBlock):
     download_image = DocumentChooserBlock(icon='image')
     director = blocks.CharBlock()
     divisions = blocks.ListBlock(BureauStructureDivision())
-    office_of_the_director = blocks.ListBlock(BureauStructureOffice(),
-                             label='Office of the Director')
+    office_of_the_director = blocks.ListBlock(
+        BureauStructureOffice(), label='Office of the Director'
+    )
+
     class Meta:
         icon = None
         template = '_includes/organisms/bureau-structure.html'
@@ -433,7 +482,7 @@ class FullWidthText(blocks.StreamBlock):
     cta = molecules.CallToAction()
     related_links = molecules.RelatedLinks()
     table = Table(editable=False)
-    table_block = AtomicTableBlock(table_options={'renderer':'html'})
+    table_block = AtomicTableBlock(table_options={'renderer': 'html'})
 
     class Meta:
         icon = 'edit'
@@ -520,8 +569,10 @@ class FilterControls(BaseExpandable):
                                 label='Filter Title')
     post_date_description = blocks.CharBlock(default='Published')
     categories = blocks.StructBlock([
-        ('filter_category', blocks.BooleanBlock(default=True, required=False)),
-        ('show_preview_categories', blocks.BooleanBlock(default=True, required=False)),
+        ('filter_category',
+         blocks.BooleanBlock(default=True, required=False)),
+        ('show_preview_categories',
+         blocks.BooleanBlock(default=True, required=False)),
         ('page_type', blocks.ChoiceBlock(choices=ref.page_types,
                                          required=False)),
     ])
@@ -570,7 +621,8 @@ class HTMLBlock(blocks.StructBlock):
         required=True,
         regex=r'^https://(s3.amazonaws.com/)?files.consumerfinance.gov/.+$',
         error_messages={
-            'required': 'The HTML URL field is required for rendering raw HTML from a remote source.',
+            'required': 'The HTML URL field is required for rendering raw '
+                        'HTML from a remote source.',
             'invalid': 'The URL is invalid or not allowed. ',
         }
     )
@@ -580,7 +632,24 @@ class HTMLBlock(blocks.StructBlock):
         resp.raise_for_status()
         return self.render_basic(resp.content, context=context)
 
-
     class Meta:
         label = 'HTML Block'
         icon = 'code'
+
+
+class ChartBlock(blocks.StructBlock):
+    element_id = blocks.CharBlock(
+        required=True,
+        label='Element ID',
+        help_text='See the element IDs in '
+        'https://github.com/cfpb/consumer-credit-trends/'
+        'blob/master/src/static/js/templates/charts.js'
+    )
+    title = blocks.CharBlock(required=False)
+    data_source = blocks.CharBlock(required=False)
+    note = blocks.CharBlock(required=False)
+
+    class Meta:
+        label = 'Chart Block'
+        icon = 'image'
+        template = '_includes/organisms/chart.html'