+++ cfgov/v1/migrations/0031_add_social_media_customization.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.db import migrations, models
+import wagtail.wagtailcore.blocks
 import wagtail.wagtailcore.fields
 import wagtail.wagtailsnippets.blocks
+from django.db import migrations, models
+
 import v1.models.snippets
-import wagtail.wagtailcore.blocks
 
 
 class Migration(migrations.Migration):