+++ cfgov/v1/models/images.py
@@ -2,12 +2,11 @@ from django.db import models
 from django.db.models.signals import pre_delete
 from django.dispatch import receiver
 from django.utils.six import string_types
-from wagtail.wagtailimages.image_operations import (
-    DoNothingOperation, MinMaxOperation, WidthHeightOperation
-)
-from wagtail.wagtailimages.models import (
-    AbstractImage, AbstractRendition, Filter, Image
-)
+from wagtail.wagtailimages.image_operations import (DoNothingOperation,
+                                                    MinMaxOperation,
+                                                    WidthHeightOperation)
+from wagtail.wagtailimages.models import (AbstractImage, AbstractRendition,
+                                          Filter, Image)
 
 
 class CFGOVImage(AbstractImage):