+++ cfgov/v1/models/sublanding_page.py
@@ -1,20 +1,22 @@
-from wagtail.wagtailcore.fields import StreamField
-from wagtail.wagtailadmin.edit_handlers import TabbedInterface, ObjectList, \
-    StreamFieldPanel
+import logging
+
+from wagtail.wagtailadmin.edit_handlers import (
+    ObjectList,
+    StreamFieldPanel,
+    TabbedInterface
+)
 from wagtail.wagtailcore import blocks
+from wagtail.wagtailcore.fields import StreamField
 from wagtail.wagtailcore.models import PageManager
 from wagtail.wagtailimages.blocks import ImageChooserBlock
 
+from jobmanager.models import JobListingList
+from v1 import blocks as v1_blocks
+from v1.atomic_elements import molecules, organisms
+from v1.forms import FilterableListForm
 from v1.models.base import CFGOVPage
 from v1.models.learn_page import AbstractFilterPage
 
-from v1 import blocks as v1_blocks
-from v1.atomic_elements import molecules, organisms
-
-from jobmanager.models import JobListingList
-from v1.forms import FilterableListForm
-
-import logging
 logger = logging.getLogger(__name__)
 
 
@@ -33,7 +35,8 @@ class SublandingPage(CFGOVPage):
         ('post_preview_snapshot', organisms.PostPreviewSnapshot()),
         ('well', organisms.Well()),
         ('table', organisms.Table(editable=False)),
-        ('table_block', organisms.AtomicTableBlock(table_options={'renderer': 'html'})),
+        ('table_block', organisms.AtomicTableBlock(
+            table_options={'renderer': 'html'})),
         ('contact', organisms.MainContactInfo()),
         ('formfield_with_button', molecules.FormFieldWithButton()),
         ('reg_comment', organisms.RegComment()),
@@ -45,11 +48,14 @@ class SublandingPage(CFGOVPage):
         ('paragraph', blocks.RichTextBlock(icon='edit')),
         ('breakout_image', blocks.StructBlock([
             ('image', ImageChooserBlock()),
-            ('is_round', blocks.BooleanBlock(required=False, default=True,
+            ('is_round', blocks.BooleanBlock(required=False,
+                                             default=True,
                                              label='Round?')),
             ('icon', blocks.CharBlock(help_text='Enter icon class name.')),
-            ('heading', blocks.CharBlock(required=False, label='Introduction Heading')),
-            ('body', blocks.TextBlock(required=False, label='Introduction Body')),
+            ('heading', blocks.CharBlock(required=False,
+                                         label='Introduction Heading')),
+            ('body', blocks.TextBlock(required=False,
+                                      label='Introduction Body')),
         ], heading='Breakout Image', icon='image')),
         ('related_posts', organisms.RelatedPosts()),
         ('job_listing_list', JobListingList()),
@@ -78,15 +84,19 @@ class SublandingPage(CFGOVPage):
 
     def get_browsefilterable_posts(self, request, limit):
         hostname = request.site.hostname
-        filter_pages = [p.specific for p in self.get_appropriate_descendants(hostname)
-                        if 'FilterablePage' in p.specific_class.__name__
-                        and 'archive' not in p.title.lower()]
+        filter_pages = [p.specific
+                        for p in self.get_appropriate_descendants(hostname)
+                        if 'FilterablePage' in p.specific_class.__name__ and
+                        'archive' not in p.title.lower()]
         posts_tuple_list = []
         for page in filter_pages:
-            base_query = AbstractFilterPage.objects.live().filter(CFGOVPage.objects.child_of_q(page))
+            base_query = AbstractFilterPage.objects.live().filter(
+                CFGOVPage.objects.child_of_q(page))
             logger.info('Filtering by parent {}'.format(page))
             form_id = str(page.form_id())
             form = FilterableListForm(hostname=hostname, base_query=base_query)
             for post in form.get_page_set():
                 posts_tuple_list.append((form_id, post))
-        return sorted(posts_tuple_list, key=lambda p: p[1].date_published, reverse=True)[:limit]
+        return sorted(posts_tuple_list,
+                      key=lambda p: p[1].date_published,
+                      reverse=True)[:limit]