+++ cfgov/v1/tests/test_filterable_list_form.py
@@ -1,5 +1,4 @@
 from django.test import TestCase
-
 from wagtail.wagtailcore.models import Site
 
 from v1.forms import FilterableListForm
@@ -9,11 +8,12 @@ from v1.models.learn_page import AbstractFilterPage, EventPage
 from v1.tests.wagtail_pages.helpers import publish_page
 from v1.util.categories import clean_categories
 
+
 class TestFilterableListForm(TestCase):
 
     def setUpFilterableForm(self, data=None):
         hostname = Site.objects.get(is_default_site=True).hostname
-        base_query = AbstractFilterPage.objects.live_shared(hostname)
+        base_query = AbstractFilterPage.objects.live()
         form = FilterableListForm(hostname=hostname, base_query=base_query)
         form.is_bound = True
         form.cleaned_data = data