+++ cfgov/v1/views/__init__.py
@@ -1,16 +1,15 @@
-from wagtail.wagtailcore.models import Page
-from django.shortcuts import render, redirect, get_object_or_404
-from django.core.exceptions import PermissionDenied, ObjectDoesNotExist
-from wagtail.wagtailadmin import messages as wagtail_messages
+from django.core.exceptions import ObjectDoesNotExist
 from django.utils.translation import ugettext as _
-from django.core.urlresolvers import reverse
 from django.conf import settings
-from django.http import HttpResponse, Http404
+from django.http import Http404
 
 from django.contrib.auth import get_user_model
 from django.contrib import messages
-from django.contrib.auth import update_session_auth_hash,\
-        REDIRECT_FIELD_NAME, login
+from django.contrib.auth import (
+    update_session_auth_hash,
+    REDIRECT_FIELD_NAME,
+    login
+)
 from django.contrib.auth.tokens import default_token_generator
 from django.contrib.auth.decorators import login_required
 from django.contrib.sites.shortcuts import get_current_site
@@ -25,16 +24,11 @@ from django.template.response import TemplateResponse
 from django.core.urlresolvers import resolve
 
 from wagtail.wagtailadmin.views import account
-from wagtail.wagtailusers.views.users import add_user_perm, change_user_perm
-from wagtail.wagtailadmin.utils import permission_required
 
 from v1.auth_forms import (
     CFGOVPasswordChangeForm, CFGOVSetPasswordForm, LoginForm
 )
-from v1.util.util import (
-    all_valid_destinations_for_request, valid_destination_for_request
-)
-
+from v1.util.util import all_valid_destinations_for_request
 
 
 # Overrided Wagtail Views
@@ -56,7 +50,10 @@ def change_password(request):
             form.save()
             update_session_auth_hash(request, form.user)
 
-            messages.success(request, _("Your password has been changed successfully!"))
+            messages.success(
+                request,
+                _("Your password has been changed successfully!")
+            )
             return redirect('wagtailadmin_account')
         else:
             if '__all__' in form.errors:
@@ -69,7 +66,6 @@ def change_password(request):
         'form': form,
         'can_change_password': can_change_password,
     })
-
 
 
 @sensitive_post_parameters()
@@ -103,10 +99,12 @@ def login_with_lockout(request, template_name='wagtailadmin/login.html'):
 
             login(request, form.get_user())
 
-            return HttpResponseRedirect('/login/check_permissions/?next=' + redirect_to)
+            return HttpResponseRedirect(
+                '/login/check_permissions/?next=' + redirect_to)
     else:
-	if request.user.is_authenticated():
-            return HttpResponseRedirect('/login/check_permissions/?next=' + redirect_to)
+        if request.user.is_authenticated():
+            return HttpResponseRedirect(
+                '/login/check_permissions/?next=' + redirect_to)
         form = LoginForm(request)
 
     current_site = get_current_site(request)
@@ -130,8 +128,9 @@ def check_permissions(request):
                                    request.GET.get(REDIRECT_FIELD_NAME, ''))
 
     if not request.user.is_authenticated():
-       return HttpResponseRedirect("%s?%s=%s" % (settings.LOGIN_URL,
-          REDIRECT_FIELD_NAME, redirect_to))
+        return HttpResponseRedirect(
+            "%s?%s=%s" % (settings.LOGIN_URL, REDIRECT_FIELD_NAME, redirect_to)
+        )
     view, args, kwargs = resolve(redirect_to)
     kwargs['request'] = request
     try:
@@ -143,18 +142,24 @@ def check_permissions(request):
         # this indicates a permissions problem
         # (there may be a better way)
         if REDIRECT_FIELD_NAME + '=' in response.url:
-            return render(request, "wagtailadmin/access_denied.html",
-            context= {'attempted_to_reach': redirect_to,
-                      'destinations': all_valid_destinations_for_request(request)})
+            return render(
+                request,
+                "wagtailadmin/access_denied.html",
+                context={
+                    'attempted_to_reach': redirect_to,
+                    'destinations': all_valid_destinations_for_request(request)
+                }
+            )
 
     return HttpResponseRedirect(redirect_to)
 
 
 @sensitive_post_parameters()
 @never_cache
-def custom_password_reset_confirm(request, uidb64=None, token=None,
-                                  template_name='wagtailadmin/account/password_reset/confirm.html',
-                                  post_reset_redirect='wagtailadmin_password_reset_complete'):
+def custom_password_reset_confirm(
+        request, uidb64=None, token=None,
+        template_name='wagtailadmin/account/password_reset/confirm.html',
+        post_reset_redirect='wagtailadmin_password_reset_complete'):
     """
     View that checks the hash in a password reset link and presents a
     form for entering a new password.
@@ -195,17 +200,22 @@ def custom_password_reset_confirm(request, uidb64=None, token=None,
 
     return TemplateResponse(request, template_name, context)
 
+
 @never_cache
 @login_required
 def welcome(request):
     valid_destinations = all_valid_destinations_for_request(request)
-
 
     if len(valid_destinations) == 1:
         redirect_to = valid_destinations[0][1]
         return HttpResponseRedirect(redirect_to)
 
     else:
-        return render(request, 'welcome.html', {'destinations': valid_destinations})
+        return render(
+            request,
+            'welcome.html',
+            {'destinations': valid_destinations}
+        )
 
-password_reset_confirm = account._wrap_password_reset_view(custom_password_reset_confirm)
+password_reset_confirm = account._wrap_password_reset_view(
+    custom_password_reset_confirm)